+++ app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/designsystem/styles/CIcons.kt
@@ -1,0 +1,12 @@
+package org.archivekeep.app.desktop.ui.designsystem.styles
+
+import compose.icons.TablerIcons
+import compose.icons.tablericons.Files
+import compose.icons.tablericons.Server
+
+class CIcons {
+    companion object {
+        val Storage = TablerIcons.Server
+        val Repository = TablerIcons.Files
+    }
+}